web-terminal/server/requirements.txt
Tornado == 4.5.2
futures
motor == 0.7
pyjwt
validictory
passlib
sockjs-tornado
momoko
pyzmq
bson
zope.interface
